Fixed unittest setup. Run "python run_tests.py" to run all tests.
Removed the tests/ directory. It's convenient to be in the same directory as the code you're importing.

run_tests.py will find any file named "test*.py" and will run the tests it contains. A "test" is defined as any method that is named like "test*(self)" in a class that inherits from unittest.TestCase.


git-svn-id: http://projecthexapod.com/svn@127 c308ba4a-1832-4cc7-b836-58aacb0a83c6